pyproject.toml
[project]
name = "trackid"
version = "0.1.0"
description = "CLI tool for identifying music tracks using Shazam and ACRCloud"
readme = "README.md"
requires-python = ">=3.11,<3.14"
authors = [
    { name = "Patric Fornasier" }
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "shazamio>=0.6.0",
    "pyacrcloud>=1.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
trackid = "trackid.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trackid"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
